Always prefer 'Tango' icon theme (#250)
The legacy behavior was to attempt to detect if an existing theme was
already set which supplied some of the needed icons, and if not, use the
Tango theme specifically. If that didn't work either, revert the change
to the original theme (or lack thereof).

The current behavior changes the theme only if no theme is set, and only
if tango_icons_vendor is installed.

This change makes qt_gui always prefer the 'Tango' theme, and always
prefer the version supplied by tango_icons_vendor (if available). If
'Tango' is not available, fall back to the system theme.

Signed-off-by: Scott K Logan <logans@cottsay.net>
Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>
Co-authored-by: Chris Lalancette <clalancette@openrobotics.org>