Skip running the tests on Windows Debug. (#292)
* Skip running the tests on Windows Debug.

When we run CI on Windows Debug, Python expects to be able
to find compiled libraries with an extra _d in their filename
that contains the debugging ABI.

However, we don't currently have a debug library available for PyQt5,
and thus when running under Windows Debug, the tests cannot succeed.
While we could build a debug library for PyQt5, that is actually a
massive undertaking. Instead, just skip these tests when we are on
Windows Debug. While that loses us a bit of coverage, it won't matter
much; we are still testing this on Windows (Release) and on Linux.

The way in which this patch determines that this is a Windows Debug
interpreter is a little chintzy. It just uses the fact that the binary
is named python_d.exe to discover that. I couldn't find another way
to tell if this was a debug interpreter, but I am open to other ideas.

* Update to using importlib.machinery.

Thanks to Shane for figuring out this way to tell
if we are in a debug interpreter.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>